requirements.txt
Django==5.0.6
djangorestframework==3.15.2
python-dotenv==1.0.1
dnspython==2.6.1
psycopg2==2.9.5
beautifulsoup4==4.12.3
requests==2.32.3
pytest-django==4.8.0